to spring boot

+++ backend/src/main/java/com/video/tanmu/param/TanmuInsertParam.java
@@ -23,6 +23,6 @@ public class TanmuInsertParam {
     /**
      * 弹幕时间
      */
-    private Integer currentTime;
+    private Double currentTime;
 
 }
